--- src/main/kotlin/org/jqassistant/tooling/intellij/plugin/editor/effectiveConfig/EffectiveConfigurationPanel.kt
@@ -1,12 +1,0 @@
-package org.jqassistant.tooling.intellij.plugin.editor.effectiveConfig
-
-import com.intellij.ui.components.JBScrollPane
-import javax.swing.JTextPane
-
-class EffectiveConfigurationPanel(config : String) : JBScrollPane() {
-    init {
-        val text = JTextPane()
-        text.text = config
-        this.setViewportView(text)
-    }
-}
