+++ src/main/kotlin/org/jqassistant/tooling/intellij/plugin/editor/effectiveConfig/EffectiveConfigurationScrollPane.kt
@@ -1,0 +1,15 @@
+package org.jqassistant.tooling.intellij.plugin.editor.effectiveConfig
+
+import com.intellij.ui.components.JBScrollPane
+import javax.swing.JTextPane
+
+class EffectiveConfigurationScrollPane() : JBScrollPane() {
+    private val textPane = JTextPane()
+
+    init {
+        this.setViewportView(textPane)
+    }
+    fun setText(config : String) {
+        this.textPane.text = config
+    }
+}
